requirements.txt
# Core dependencies
numpy>=1.24
rasterio>=1.3
Pillow>=9.0
matplotlib>=3.5
requests>=2.28
tqdm>=4.60

# HF / STAC / utilities
datasets
huggingface-hub
planetary-computer
pystac-client
shapely
opencv-python
timm
torchgeo

# NOTE: Install PyTorch separately with the correct CUDA/CPU wheel per https://pytorch.org